+++ crates/vm/backends/revm/db.rs
@@ -1,13 +1,20 @@
-use ethrex_common::{types::BlockHash, Address as CoreAddress, H256 as CoreH256};
-use ethrex_storage::{error::StoreError, Store};
+use std::collections::HashMap;
+
+use ethrex_common::{
+    types::{Block, BlockHash},
+    Address as CoreAddress, H256 as CoreH256,
+};
+use ethrex_storage::{error::StoreError, hash_address, hash_key, Store};
+use ethrex_trie::{NodeRLP, TrieError};
 use revm::primitives::{
     AccountInfo as RevmAccountInfo, Address as RevmAddress, Bytecode as RevmBytecode,
     Bytes as RevmBytes, B256 as RevmB256, U256 as RevmU256,
 };
 
 use crate::{
-    db::{ExecutionDB, StoreWrapper},
-    ChainConfig, EvmError,
+    db::{get_potential_child_nodes, ExecutionDB, StoreWrapper},
+    errors::ExecutionDBError,
+    spec_id, ChainConfig, EvmError,
 };
 
 /// State used when running the EVM. The state can be represented with a [StoreWrapper] database, or
@@ -155,5 +162,187 @@ impl revm::DatabaseRef for StoreWrapper {
             .get_block_header(number)?
             .map(|header| RevmB256::from_slice(&header.compute_block_hash().0))
             .ok_or_else(|| StoreError::Custom(format!("Block {number} not found")))
+    }
+}
+impl StoreWrapper {
+    pub fn to_exec_db_revm(&self, block: &Block) -> Result<ExecutionDB, ExecutionDBError> {
+        // TODO: Simplify this function and potentially merge with the implementation for
+        // RpcDB.
+
+        let parent_hash = block.header.parent_hash;
+        let chain_config = self.store.get_chain_config()?;
+
+        // pre-execute and get all state changes
+        let cache = ExecutionDB::pre_execute_revm(
+            block,
+            chain_config.chain_id,
+            spec_id(&chain_config, block.header.timestamp),
+            self,
+        )
+        .map_err(|err| Box::new(EvmError::from(err)))?; // TODO: ugly error handling
+        let store_wrapper = cache.db;
+
+        // index read and touched account addresses and storage keys
+        let index = cache.accounts.iter().map(|(address, account)| {
+            let address = CoreAddress::from(address.0.as_ref());
+            let storage_keys: Vec<_> = account
+                .storage
+                .keys()
+                .map(|key| CoreH256::from_slice(&key.to_be_bytes_vec()))
+                .collect();
+            (address, storage_keys)
+        });
+
+        // fetch all read/written values from store
+        let cache_accounts = cache.accounts.iter().filter_map(|(address, account)| {
+            let address = CoreAddress::from(address.0.as_ref());
+            // filter new accounts (accounts that didn't exist before) assuming our store is
+            // correct (based on the success of the pre-execution).
+            if store_wrapper
+                .store
+                .get_account_info_by_hash(parent_hash, address)
+                .is_ok_and(|account| account.is_some())
+            {
+                Some((address, account))
+            } else {
+                None
+            }
+        });
+        let accounts = cache_accounts
+            .clone()
+            .map(|(address, _)| {
+                // return error if account is missing
+                let account = match store_wrapper
+                    .store
+                    .get_account_info_by_hash(parent_hash, address)
+                {
+                    Ok(Some(some)) => Ok(some),
+                    Err(err) => Err(ExecutionDBError::Store(err)),
+                    Ok(None) => unreachable!(), // we are filtering out accounts that are not present
+                                                // in the store
+                };
+                Ok((address, account?))
+            })
+            .collect::<Result<HashMap<_, _>, ExecutionDBError>>()?;
+        let code = cache_accounts
+            .clone()
+            .map(|(_, account)| {
+                // return error if code is missing
+                let hash = CoreH256::from(account.info.code_hash.0);
+                Ok((
+                    hash,
+                    store_wrapper
+                        .store
+                        .get_account_code(hash)?
+                        .ok_or(ExecutionDBError::NewMissingCode(hash))?,
+                ))
+            })
+            .collect::<Result<_, ExecutionDBError>>()?;
+        let storage = cache_accounts
+            .map(|(address, account)| {
+                // return error if storage is missing
+                Ok((
+                    address,
+                    account
+                        .storage
+                        .keys()
+                        .map(|key| {
+                            let key = CoreH256::from(key.to_be_bytes());
+                            let value = store_wrapper
+                                .store
+                                .get_storage_at_hash(parent_hash, address, key)
+                                .map_err(ExecutionDBError::Store)?
+                                .ok_or(ExecutionDBError::NewMissingStorage(address, key))?;
+                            Ok((key, value))
+                        })
+                        .collect::<Result<HashMap<_, _>, ExecutionDBError>>()?,
+                ))
+            })
+            .collect::<Result<HashMap<_, _>, ExecutionDBError>>()?;
+        let block_hashes = cache
+            .block_hashes
+            .into_iter()
+            .map(|(num, hash)| (num.try_into().unwrap(), CoreH256::from(hash.0)))
+            .collect();
+        // WARN: unwrapping because revm wraps a u64 as a U256
+
+        // get account proofs
+        let state_trie = self
+            .store
+            .state_trie(block.hash())?
+            .ok_or(ExecutionDBError::NewMissingStateTrie(parent_hash))?;
+        let parent_state_trie = self
+            .store
+            .state_trie(parent_hash)?
+            .ok_or(ExecutionDBError::NewMissingStateTrie(parent_hash))?;
+        let hashed_addresses: Vec<_> = index
+            .clone()
+            .map(|(address, _)| hash_address(&address))
+            .collect();
+        let initial_state_proofs = parent_state_trie.get_proofs(&hashed_addresses)?;
+        if let Some(a) = &initial_state_proofs.0 {
+            println!("a: {}", a.len());
+        }
+        let final_state_proofs: Vec<_> = hashed_addresses
+            .iter()
+            .map(|hashed_address| Ok((hashed_address, state_trie.get_proof(hashed_address)?)))
+            .collect::<Result<_, TrieError>>()?;
+        let potential_account_child_nodes = final_state_proofs
+            .iter()
+            .filter_map(|(hashed_address, proof)| get_potential_child_nodes(proof, hashed_address))
+            .flat_map(|nodes| nodes.into_iter().map(|node| node.encode_raw()))
+            .collect();
+        let state_proofs = (
+            initial_state_proofs.0,
+            [initial_state_proofs.1, potential_account_child_nodes].concat(),
+        );
+
+        // get storage proofs
+        let mut storage_proofs = HashMap::new();
+        let mut final_storage_proofs = HashMap::new();
+        for (address, storage_keys) in index {
+            let Some(parent_storage_trie) = self.store.storage_trie(parent_hash, address)? else {
+                // the storage of this account was empty or the account is newly created, either
+                // way the storage trie was initially empty so there aren't any proofs to add.
+                continue;
+            };
+            let storage_trie = self.store.storage_trie(block.hash(), address)?.ok_or(
+                ExecutionDBError::NewMissingStorageTrie(block.hash(), address),
+            )?;
+            let paths = storage_keys.iter().map(hash_key).collect::<Vec<_>>();
+
+            let initial_proofs = parent_storage_trie.get_proofs(&paths)?;
+            let final_proofs: Vec<(_, Vec<_>)> = storage_keys
+                .iter()
+                .map(|key| {
+                    let hashed_key = hash_key(key);
+                    let proof = storage_trie.get_proof(&hashed_key)?;
+                    Ok((hashed_key, proof))
+                })
+                .collect::<Result<_, TrieError>>()?;
+
+            let potential_child_nodes: Vec<NodeRLP> = final_proofs
+                .iter()
+                .filter_map(|(hashed_key, proof)| get_potential_child_nodes(proof, hashed_key))
+                .flat_map(|nodes| nodes.into_iter().map(|node| node.encode_raw()))
+                .collect();
+            let proofs = (
+                initial_proofs.0,
+                [initial_proofs.1, potential_child_nodes].concat(),
+            );
+
+            storage_proofs.insert(address, proofs);
+            final_storage_proofs.insert(address, final_proofs);
+        }
+
+        Ok(ExecutionDB {
+            accounts,
+            code,
+            storage,
+            block_hashes,
+            chain_config,
+            state_proofs,
+            storage_proofs,
+        })
     }
 }
